refactor: plaid component refactor
